Adds concurrent backend and frontend dev env
Configures the frontend development environment to run concurrently
with the backend using `concurrently`.

This allows for simultaneous development and hot-reloading of both the
frontend and backend, improving the development workflow.

Creates a `.gitignore` in the build directory.

Initializes an `.air.toml` configuration file for the backend.

To use, cd to ottofe and npm run dev.